# Fix: Scrapy spider 3x slower after implementing this